Split widget: effect row
